Init esqlVariables from local storage tabs
