+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/index.ts
@@ -94,7 +94,13 @@ export {
   useAdHocDataViews,
 } from './runtime_state';
 
-export { type TabActionInjector, createTabActionInjector, createTabItem } from './utils';
+export {
+  type TabActionInjector,
+  createTabActionInjector,
+  createTabItem,
+  parseControlGroupJson,
+  extractEsqlVariables,
+} from './utils';
 
 export {
   fromSavedObjectTabToTabState,
