+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/utils.ts
@@ -12,6 +12,9 @@ import { i18n } from '@kbn/i18n';
 import { getNextTabNumber, type TabItem } from '@kbn/unified-tabs';
 import { createAsyncThunk, miniSerializeError } from '@reduxjs/toolkit';
 import { SavedObjectNotFound } from '@kbn/kibana-utils-plugin/common';
+import type { ControlPanelsState } from '@kbn/controls-plugin/public';
+import type { ESQLControlState, ESQLControlVariable } from '@kbn/esql-types';
+import { ESQL_CONTROL } from '@kbn/controls-constants';
 import type { DiscoverInternalState, TabState } from './types';
 import type {
   InternalStateDispatch,
@@ -66,4 +69,53 @@ export const createTabItem = (allTabs: TabState[]): TabItem => {
   const label = nextNumber ? `${baseLabel} ${nextNumber}` : baseLabel;
 
   return { id, label };
+};
+
+/**
+ * Parses a JSON string into a ControlPanelsState object.
+ * If the JSON is invalid or null, it returns an empty object.
+ *
+ * @param jsonString - The JSON string to parse.
+ * @returns A ControlPanelsState object or an empty object if parsing fails.
+ */
+
+export const parseControlGroupJson = (
+  jsonString?: string | null
+): ControlPanelsState<ESQLControlState> => {
+  try {
+    return jsonString ? JSON.parse(jsonString) : {};
+  } catch (e) {
+    return {};
+  }
+};
+
+/**
+ * @param panels - The control panels state, which may be null.
+ * @description Extracts ESQL variables from the control panels state.
+ * Each ESQL control panel is expected to have a `variableName`, `variableType`, and `selectedOptions`.
+ * Returns an array of `ESQLControlVariable` objects.
+ * If `panels` is null or empty, it returns an empty array.
+ * @returns An array of ESQLControlVariable objects.
+ */
+export const extractEsqlVariables = (
+  panels: ControlPanelsState<ESQLControlState> | null
+): ESQLControlVariable[] => {
+  if (!panels || Object.keys(panels).length === 0) {
+    return [];
+  }
+  const variables = Object.values(panels).reduce((acc: ESQLControlVariable[], panel) => {
+    if (panel.type === ESQL_CONTROL) {
+      acc.push({
+        key: panel.variableName,
+        type: panel.variableType,
+        // If the selected option is not a number, keep it as a string
+        value: isNaN(Number(panel.selectedOptions?.[0]))
+          ? panel.selectedOptions?.[0]
+          : Number(panel.selectedOptions?.[0]),
+      });
+    }
+    return acc;
+  }, []);
+
+  return variables;
 };
